# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/settings": {
            "get": {
                "description": "配置相接口",
                "tags": [
                    "配置相接口"
                ],
                "summary": "获取配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": []}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "配置相接口",
                "tags": [
                    "配置相接口"
                ],
                "summary": "修改配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "配置信息",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.SettingsUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"msg\": \"更新成功\", \"data\": nil}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/settings/logo": {
            "get": {
                "description": "配置相接口",
                "tags": [
                    "配置相接口"
                ],
                "summary": "获取 Logo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"path\": logoPreview}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/settings/logoUpload": {
            "post": {
                "description": "配置相接口",
                "tags": [
                    "配置相接口"
                ],
                "summary": "上传 Logo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Logo",
                        "name": "logo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"path\": logoPreview}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "service.SettingsUpdate": {
            "type": "object",
            "properties": {
                "externalUrl": {
                    "type": "string"
                },
                "issuer": {
                    "type": "string"
                },
                "ldapAddress": {
                    "type": "string"
                },
                "ldapBindDn": {
                    "type": "string"
                },
                "ldapBindPassword": {
                    "type": "string"
                },
                "ldapFilterAttribute": {
                    "type": "string"
                },
                "ldapSearchDn": {
                    "type": "string"
                },
                "ldapUserPasswordExpireDate": {
                    "type": "string"
                },
                "mfa": {
                    "type": "string"
                }
            }
        },
...
    }
}